Separate mysql/time (#3276)
